pyproject.toml
[project]
name = "get-file"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
